requirements.txt
addict==2.4.0
aiohttp==3.12.15
anyio==4.9.0
beartype==0.14.1
click==8.2.2
cryptography==45.0.5
curl_cffi==0.12.0
dashscope
datasets==3.6.0
einops==0.8.1
face_alignment==1.4.1
fastapi==0.116.1
fastcluster==1.3.0
ffmpy==0.6.1
filelock==3.18.0
funasr==1.2.6
gradio
g4f
hdbscan==0.8.40
huggingface-hub==0.34.3
imageio==2.37.0
imageio-ffmpeg==0.6.0
kaldiio==2.18.1
joblib==1.5.1
librosa==0.11.0
matplotlib
ml_collections==0.1.1
modelscope==1.28.1
moviepy==1.0.3
numba==0.61.2
numpy==2.2.6
onnxruntime==1.19.2
openai
openai_whisper==20250625
opencc_python_reimplemented==0.1.7
opencv_python==4.12.0.88
packaging
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
pyannote.audio
pypinyin==0.44.0
python_speech_features==0.6
python-Levenshtein
PyYAML==6.0.2
requests
rotary_embedding_torch==0.8.9
scikit-learn==1.7.1
scipy==1.15.3
simplejson
soundfile==0.13.1
starlette==0.47.2
tensorboardX==2.6.4
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers
umap_learn==0.5.7
